agent/core: Use internalState for (*State).Dump()
Step 2 of 2. The general idea is that we already have golangci-lint
checking for exhaustive struct field initialization, so we might as well
use that to make sure that our state dumping is up to date, rather than
having a comment saying "update dumpstate when you change something!"

While this still requires manually implementing deep copying, it should
at least be nicer than manually copying the types as well.